+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fixed models with `XLapJoint` fail to serialize.
+* Print warning when CategoryRules are inspecting elements which have no `category` attributes set.
 
 ### Removed